refactor: combine meta helpers
