+++ src/index.js
@@ -2,7 +2,7 @@ import { version } from '../package.json'
 import createMixin from './shared/mixin'
 import setOptions from './shared/options'
 import $meta from './server/$meta'
-import hasMetaInfo from './shared/hasMetaInfo'
+import { hasMetaInfo } from './shared/meta-helpers'
 
 /**
  * Plugin install function.
