--- src/shared/hasMetaInfo.js
@@ -1,6 +1,0 @@
-import { isObject } from './is-type'
-
-// Vue $root instance has a _vueMeta object property, otherwise its a boolean true
-export default function hasMetaInfo(vm = this) {
-  return vm && (vm._vueMeta === true || isObject(vm._vueMeta))
-}
