+++ test/unit/getComponentOptions.test.js
@@ -1,5 +1,5 @@
 import getComponentOption from '../../src/shared/getComponentOption'
-import inMetaInfoBranch from '../../src/shared/inMetaInfoBranch'
+import { inMetaInfoBranch } from '../../src/shared/meta-helpers'
 import { mount, getVue, loadVueMetaPlugin } from '../utils'
 
 describe('getComponentOption', () => {
